Simplified vtctld.Explorer code.
No need to have the interface now that we only  have one implementation.
Moved everything into vtctld/explorer.go.
