+++ go/cmd/vtctld/plugin_consultopo.go
@@ -16,19 +16,8 @@ limitations under the License.
 
 package main
 
-// Imports and register the 'consul' topo.Server and its Explorer.
+// Imports and register the 'consul' topo.Server.
 
 import (
-	"github.com/youtube/vitess/go/vt/servenv"
-	"github.com/youtube/vitess/go/vt/topo/consultopo"
-	"github.com/youtube/vitess/go/vt/vtctld"
+	_ "github.com/youtube/vitess/go/vt/topo/consultopo"
 )
-
-func init() {
-	// Wait until flags are parsed, so we can check which topo server is in use.
-	servenv.OnRun(func() {
-		if s, ok := ts.Impl.(*consultopo.Server); ok {
-			vtctld.HandleExplorer("consul", vtctld.NewBackendExplorer(s))
-		}
-	})
-}
