+++ go/cmd/vtctld/plugin_etcd2topo.go
@@ -16,19 +16,8 @@ limitations under the License.
 
 package main
 
-// Imports and register the 'etcd2' topo.Server and its Explorer.
+// Imports and register the 'etcd2' topo.Server.
 
 import (
-	"github.com/youtube/vitess/go/vt/servenv"
-	"github.com/youtube/vitess/go/vt/topo/etcd2topo"
-	"github.com/youtube/vitess/go/vt/vtctld"
+	_ "github.com/youtube/vitess/go/vt/topo/etcd2topo"
 )
-
-func init() {
-	// Wait until flags are parsed, so we can check which topo server is in use.
-	servenv.OnRun(func() {
-		if s, ok := ts.Impl.(*etcd2topo.Server); ok {
-			vtctld.HandleExplorer("etcd2", vtctld.NewBackendExplorer(s))
-		}
-	})
-}
