--- go/vt/vtctld/explorer/explorer.go
@@ -1,32 +1,0 @@
-/*
-Copyright 2017 Google Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreedto in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package explorer
-
-import "net/http"
-
-// Result is what the explorer returns. It represents one directory node.
-type Result struct {
-	Data     string
-	Children []string
-	Error    string
-}
-
-// Explorer allows exploring a topology server.
-type Explorer interface {
-	// HandlePath returns a Result for the given path.
-	HandlePath(url string, r *http.Request) *Result
-}
